chore: Put code into module and add documentation
chore: Adjust licensing info
