+++ src/prometheus_fastapi_instrumentator/middleware.py
@@ -8,54 +8,11 @@ from prometheus_client import Gauge
 from starlette.datastructures import Headers
 from starlette.requests import Request
 from starlette.responses import Response
-from starlette.routing import Match, Mount
 
-from prometheus_fastapi_instrumentator import metrics
+from prometheus_fastapi_instrumentator import metrics, routing
 
 if TYPE_CHECKING:
     from asgiref.typing import ASGISendEvent
-
-
-def _get_route_name(scope, routes, route_name=None) -> Optional[str]:
-    for route in routes:
-        match, child_scope = route.matches(scope)
-        if match == Match.FULL:
-            route_name = route.path
-            child_scope = {**scope, **child_scope}
-            if isinstance(route, Mount) and route.routes:
-                child_route_name = _get_route_name(child_scope, route.routes, route_name)
-                if child_route_name is None:
-                    route_name = None
-                else:
-                    route_name += child_route_name
-            return route_name
-        elif match == Match.PARTIAL and route_name is None:
-            route_name = route.path
-    return None
-
-
-def get_route_name(request: Request) -> Optional[str]:
-    app = request.app
-    scope = request.scope
-    routes = app.routes
-    route_name = _get_route_name(scope, routes)
-
-    # Starlette magically redirects requests if the path matches a route name with a trailing slash
-    # appended or removed. To not spam the transaction names list, we do the same here and put these
-    # redirects all in the same "redirect trailing slashes" transaction name
-    if not route_name and app.router.redirect_slashes and scope["path"] != "/":
-        redirect_scope = dict(scope)
-        if scope["path"].endswith("/"):
-            redirect_scope["path"] = scope["path"][:-1]
-            trim = True
-        else:
-            redirect_scope["path"] = scope["path"] + "/"
-            trim = False
-
-        route_name = _get_route_name(redirect_scope, routes)
-        if route_name is not None:
-            route_name = route_name + "/" if trim else route_name[:-1]
-    return route_name
 
 
 class PrometheusInstrumentatorMiddleware:
@@ -184,7 +141,7 @@ class PrometheusInstrumentatorMiddleware:
                 template or if no template the path. Second element tells you
                 if the path is templated or not.
         """
-        route_name = get_route_name(request)
+        route_name = routing.get_route_name(request)
         return route_name or request.url.path, True if route_name else False
 
     def _is_handler_excluded(self, handler: str, is_templated: bool) -> bool:
